requirements.txt
Django==1.8
django-braces==1.4.0
django-compressor==1.4
django-debug-toolbar==1.3.0
django-extensions==1.5.2
celery==3.1.17
gunicorn==19.3.0
ipdb==0.8
ipython==3.1.0
psycopg2==2.6
Werkzeug>=0.10.4
youtube-dl>=2014.06.26

